llm_analysis/llm_analysis/requirements.txt
fastapi 
uvicorn 
gitpython 
pydantic 
requests 
networkx
GitPython
python-dotenv
SQLAlchemy
sqlmodel
psycopg2-binary>=2.9
yandex-cloud-ml-sdk